+++ plugins/main/server/plugin.ts
@@ -191,6 +191,10 @@ export class WazuhPlugin implements Plugin<WazuhPluginSetup, WazuhPluginStart> {
     return !!plugins.notificationsDashboards;
   }
 
+  isAlertingDashboardsAvailable(plugins: PluginSetup): boolean {
+    return !!plugins.alertingDashboards;
+  }
+
   public async setup(core: CoreSetup, plugins: PluginSetup) {
     this.logger.debug('Wazuh-wui: Setup');
 
@@ -229,7 +233,10 @@ export class WazuhPlugin implements Plugin<WazuhPluginSetup, WazuhPluginStart> {
     // Detect Notifications plugin availability to conditionally register tasks
     if (this.isNotificationsDashboardsAvailable(plugins)) {
       core.healthCheck.register(
-        initializeDefaultNotificationChannel(notificationClient),
+        initializeDefaultNotificationChannel(
+          notificationClient,
+          this.isAlertingDashboardsAvailable(plugins),
+        ),
       );
     } else {
       this.logger.debug(